Fix: rename component from my/component to my-component
