+++ src/app/components/003-FeatureOverviewSection/index.tsx
@@ -1,4 +1,4 @@
-import ReusableTextImage from "../../006-Resusable-Text-Image";
+import ReusableTextImage from "../006-Resusable-Text-Image";
 
 const OverviewSection = () => {
     return (
